Squashing some changes
