--- database/routes/src/main/java/ua/vertex/waypoint/DAO/WayPointDAO.java
@@ -1,16 +1,0 @@
-package ua.vertex.waypoint.DAO;
-
-import ua.vertex.waypoint.Entity.WayPoint;
-
-import java.util.List;
-
-/**
- * Created by Vasyl on 18/05/2016.
- */
-public interface WayPointDAO {
-    void create(WayPoint wayPoint);
-    WayPoint read(long id);
-    void update(long id, WayPoint wayPoint);
-    void delete(long id);
-    List<WayPoint> getSortedWayPointsForRoute(long routeId);
-}
