--- database/routes/src/main/resources/embedded/insert.sql
@@ -1,7 +1,0 @@
-INSERT INTO routes.route (id, name) VALUES (1, 'first');
-INSERT INTO routes.route (id, name) VALUES (3, 'second');
-
-INSERT INTO routes.waypoint (id, track_id, x, y, height, accuracy, get_time) VALUES (1, 1, 10, 25, 400, 5, now());
-INSERT INTO routes.waypoint (id, track_id, x, y, height, accuracy, get_time) VALUES (2, 1, 50, 50, 405, 5, now());
-INSERT INTO routes.waypoint (id, track_id, x, y, height, accuracy, get_time) VALUES (3, 3, 100, 250, 600, 3, now());
-INSERT INTO routes.waypoint (id, track_id, x, y, height, accuracy, get_time) VALUES (4, 3, 110, 225, 700, 3, now());
