+++ database/users/src/main/resources/liquibase/v.1.1/db.changelog-users.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<databaseChangeLog
+        xmlns="http://www.liquibase.org/xml/ns/dbchangelog/1.9"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog/1.9
+                      http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-1.9.xsd">
+
+    <changeSet author="Vasyl" id="changelog-1.0">
+        <insert schemaName="users" tableName="payment">
+            <column name="id">1</column>
+            <column name="user_id">1</column>
+            <column name="amount">1000</column>
+        </insert>
+    </changeSet>
+
+</databaseChangeLog>
