+++ service/src/main/java/ua/vertex/config/Application.java
@@ -1,0 +1,14 @@
+package ua.vertex.config;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.liquibase.LiquibaseAutoConfiguration;
+import org.springframework.context.ApplicationContext;
+
+
+@SpringBootApplication(exclude = LiquibaseAutoConfiguration.class)
+public class Application {
+    public static void main(String[] args) {
+        ApplicationContext context = SpringApplication.run(Application.class, args);
+    }
+}
