+++ service/src/main/java/ua/vertex/dao/waypoint/WayPointDAO.java
@@ -1,0 +1,16 @@
+package ua.vertex.dao.waypoint;
+
+import entity.WayPoint;
+
+import java.util.List;
+
+/**
+ * Created by Vasyl on 18/05/2016.
+ */
+public interface WayPointDAO {
+    void create(WayPoint wayPoint);
+    WayPoint read(long id);
+    void update(long id, WayPoint wayPoint);
+    void delete(long id);
+    List<WayPoint> getSortedWayPointsForRoute(long routeId);
+}
