+++ service/src/main/java/ua/vertex/exception/RouteNotFoundException.java
@@ -1,0 +1,11 @@
+package ua.vertex.exception;
+
+public class RouteNotFoundException extends RuntimeException {
+
+    public RouteNotFoundException() {
+    }
+
+    public RouteNotFoundException(String message) {
+        super(message);
+    }
+}
