--- web/src/main/java/ua/vertex/academy/config/Configuration.java
@@ -1,13 +1,0 @@
-package ua.vertex.academy.config;
-
-import org.springframework.context.annotation.ComponentScan;
-
-
-/**
- * Created by RASTA on 19.05.2016.
- */
-@org.springframework.context.annotation.Configuration
-@ComponentScan("ua.vertex")
-public class Configuration {
-
-}
