--- web/src/main/java/ua/vertex/academy/service/RouteService.java
@@ -1,16 +1,0 @@
-package ua.vertex.academy.service;
-
-import ua.vertex.route.Entity.Route;
-
-/**
- * Created by RASTA on 19.05.2016.
- */
-public interface RouteService {
-    long create(Route route);
-
-    Route read(long id);
-
-    void update(long id, Route route);
-
-    void delete(long id);
-}
